requirements.txt
# Core Python
python-dotenv
pydantic

# LangChain (modern)
langchain
langchain-core
langchain-community
langgraph

# Hugging Face Inference
huggingface-hub
transformers
torch
accelerate
langchain-huggingface

# Groq (fallback LLM)
langchain-groq
groq

# Web framework (later use)
fastapi
uvicorn
flask
flask-cors
flask-socketio
python-socketio

# Utilities
requests
